Prefer TRUNCATE to DELETE FROM in sql calls
TRUNCATE is a more efficient way of deleting all rows in a table, so we
prefer this to DELETE FROM where we want to unconditionally remove all
the rows. Also cleanup the code to pass a pep8 lint.

Add a README to instruct users on how to setup this assignment.